+++ sample/Source/SentryPlaygroundEditor.Target.cs
@@ -10,6 +10,10 @@ public class SentryPlaygroundEditorTarget : TargetRules
 		Type = TargetType.Editor;
 		DefaultBuildSettings = BuildSettingsVersion.Latest;
 
+		// Disable Unity build and PCH files to catch missing include errors in CI
+		bUseUnityBuild = false;
+		bUsePCHFiles = false;
+
 #if UE_5_1_OR_LATER
 		IncludeOrderVersion = EngineIncludeOrderVersion.Latest;
 #endif